Removed ToDo 3 - Fixed the generated directions
Due to the fix all generated rooms have at least one exit. Because all
connections will be regenerated after succesfully changing the room it
is no problem if the only connection goes straight back into just the
room you are currently in. That could be considered as a minor bug, but
it does no harm, so it just stays in for the sake of simplicity.